Initial commit
This is the initial commit of Intel(R) Enclosure LED Utilities.
The utilities are designed primarily to be used on storage servers
utilizing MD devices (aka Linux Software RAID) for RAID arrays.

An enclosure and a disk cage has 2 or 3 leds associated with each
slot. The utilities use the leds to visualize the state of arrays
or to visually identify a slot. The ledmon does all the operations
automatically and ledctl is for manual control of leds. However
if ledmon is running it has priority over the ledctl application.

The ledctl and ledmon applications use SGPIO interface, SES-2 and
SMP protocols to control leds of an enclosure or a disk cage.
The tools support Intel(R) AHCI controller and Intel(R) SAS HBAs.
The SES-2 protocol is used for SCSI or SAS enclosures. The SGPIO
interface is used for direct-attach systems (AHCI and SAS HBA).

The applications may not work with hardware of other vendors then
Intel(R).

Signed-off-by: Artur Wojcik <artur.wojcik@intel.com>