raw bitstream (TX_GP) support
The upstream kernel has supported TX_GP writes since v3.2.  Convert
ledctl/ledmon to that interface to preclude the need for additional sgpio
support in the kernel.  Ideally the kernel would move to exporting a ses
interface for sgpio, but for now TX_GP is sufficient.

Previously sgpio read support was being used as a way to communicate the
current state of the sgpio registers, but since there is no exclusion
between updaters such read-modfiy-write cycle could not make any
consistency guarantees.  Just dump the bitstream requested by the user
and if userspace races with itself that's not the kernel's problem.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>