turn all unset LEDs off
Turning on some LEDs of one host of a controller turns off
all other LEDs of the same host of the controller,
but does not change the state of all LEDs of other hosts
and other controllers.

This patch corrects this inconsistency. Now turning on
some LEDs of one host of a controller turns off
all other LEDs of all controllers.

Signed-off-by: Lukasz Dorau <lukasz.dorau@intel.com>