Use VMD device as NVMe disk controller
Currently ledmon uses part of NVMe block device path in sysfs. Because of
this when NVMe is being hot removed from platform its controller is also
gone, so when trying to restore this drive, ledmon is not able to match
drive to controller. To fix this, pci_slot has been added to block_device
structure to chain block device with something which remains in system
after removing drive. This can be done in simpler way by using VMD device
as NVMe block device controller.

Signed-off-by: Mariusz Dabrowski <mariusz.dabrowski@intel.com>