ses: fix wrong device->encl_index assignment
The Additional Element Status descriptor can have two formats, depending
on the EIP bit status. If the bit is set, the descriptor contains the
Element Index field. Otherwise, assume that the element index on the
Additional Element Status descriptor list is the same as on the
Status/Control descriptor list.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>