ses: improve reading ses_pages
Represent ses pages with a new struct ses_page containing a fixed-length
buffer and the actual page length. This makes allocations cleaner and
makes it possible to pass a page as a function argument. Use this to
replace redundant get_page functions with one generic function. Retry up
to 3 times if sg_ll_receive_diag() fails.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>