ses: improve modifying device slot control elements
Don't clear the whole descriptors list, only modify the element
representing the block device. This way we can preserve the state of
some bits from the Enclosure Status diagnostic page.

Use a struct for the (Array) Device Slot Control Element and an enum for
element types to improve readability and enforce type checking.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>