Remove str_cat() function
Bounds checking with strcat(), or any derivative of that function, is just
too hard get right.  It is better to avoid issue of buffer overflows all
together and use snprintf() instead.  An additional upside of snprintf() is
that it is easy to see how various strings are constructed.

Add macros in pidfile path resolving, don't repeat the same strings.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>
Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@intel.com>
Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>