Add support for AMD SGPIO enclosure management (#23)
* Split the check for the vendor id outside of _is_ahci_cntrl(). This
creates a new _is_intel_ahci_cntrl() routine that validates this is
an ahci device and that the vendor id is Intel.

This is a preliminary step to allow checking for ahci devices from
vendors other than Intel.

* Update the ledmon/ledctl commands to support AMD SGPIO devices.
This is done by adding a new controller type, CNTRL_TYPE_AMD_SGPIO,
and the needed checks and callouts for this controller type in cntrl.c
and block.c.

The AMD device device specific code is added in the file src/amd_sgpio.c.
This includes the required routines to check for enclosure management
support, writing the SGPIO registers to set IBPI patterns, and getting
the sysfs enclosure management path.

The new file also handles the AMD specific code for SGPIO register
handling which include the following;

- AMD systems use an additional vendor specific SGPIO register,
referenced n the code as the amd register with register type value 0xC0.
This additional register is used to specify which set of drives the
register settings in the transmit register refer to. AMD systems control
drives insets of 8 and the amd register has an initiator bit such that
a value of 1 refers to drives 0-3, and a initiator value of 0 refers
to drives 4-7.

- Definitions for all SGPIO registers and IBPI patterns for AMD systems.

- AMD systems follow the standard SGPIO register specification for
writing registers, but not for register reads. This renders the generic
Linux driver unable to read the SGPIO registers. To resolve this a
register cache is kept of current SGPIO register states to prevent
updates of one LED setting undoing a previous LED setting on a
different drive.

Signed-off-by: Nathan Fontenot <nafonten@amd.com>