Create generic AMD interface for LED control
In preparation to support IPMI in addition to SGPIO for AMD platforms,
this patch creates a generic front-end for AMD systems that will
invoke the proper LED controls (SGPIO or IPMI) for the platform it is
running on. This will allow ledctl/ledmon to run on AMD platforms
without needing to know which interface (SGPIO or IPMI) the
platform supports. Future updates will add IPMI support.

There is no functional change in this patch. This creates a generic
front end in amd.c and leaves the SGPIO specific code in amd_sgpio.c.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>