Update ibpi pattern for block devices added by udev with previous unknown ibpi pattern
Update udev events handler in ledmon so block devices that were just added,
and which previous IBPI pattern is unknown, are correctly processed.
IBPI pattern for them could be set to Unknown in a case when several devices were
removed/added simultaneously. In a such case udev handler processes first ADD
event and for the rest of the block devices IBPI became set to UNKNOWN.