Default to SGPIO for AMD systems
A recent update to add support for IPMI on AMD systems changed
the behavior when checking for enclosure management enablement.
Previously any AMD system that returned success from
_amd_sgpio_em_enabled() would use SGPIO. The update added an explicit
check for DMI system ids that broke this behavior.

This patch corrects that behavior by defaulting to SGPIO for
AMD systems and only using the DMI id to check for systems
known to support IPMI.

As part of this change enum amd_platform was changed to enum
amd_ipmi_platforms as we only need to track systems for IPMI.
This is needed so we can determine the correct channel and slave
address for IPMI commands.

Fixes: 2a66a3c37bbf ("Add platform check for AMD systems")

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>