Check the white/blacklist earlier in discovery.
For certain controller types, enclosure management may need to be
explicitly tested and failures may be interesting errors worthy of being
logged.  However, on systems where such failures are expected, it is
beneficial to allow the whitelist and blacklist in ledmon.conf(5) to
take precedence over such tests in order to suppress repeated errors
from ledmon.