ses: change the way enclosure slots are assigned
It turns out that using enclosure slot numbers from sysfs is unreliable,
because those numbers not always correspond to element index on the
Status/Control pages. So again, always use Page 10 for that.

Get a list of slots with their addresses during enclosure discovery and
then use that when discovering block devices. Match the slot and block
device based only on sas address.

Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>