Update NPEM wait command (#117)
* npem: merge npem_write() and npem_set_slot()

The code is similar so we can call set_slot for _write.
Use _set_slot then but left _write specific checks.
Following differences were determined and resolved:

- _write returned errno based return codes.
Since it is not verified, no need to translate status codes. FIXME added.

- _write translated not supported patterns by controller to default.
Remove this functionality. Verify if pattern is supported in _set_slot.
Fail request and throw error otherwise.

* npem: Adapt npem_wait_command()

The function doesn't need to wait after imposing command. It should
check if controller is able to proceed request before issuing it.
status_cc flag should not be cleared because it means that controller is
busy.

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>