ledmon: Define ONESHOT_NORMAL for VMD
This pattern is used by ledmon internal logic. Must be defined.
Add missing definition in VMD.
Make logging less verbose because ledmon may set patterns which are not
supported- it is unnecessary to log it every time.

This will hide those logs in ledctl with default error log level
too.

Change-Id: Ice35c3d0f904d0d6df8e14fe331801dbf4ee1456
Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>