Add simple ledctl test
This is a simple letctl test which tests the slot syntax and
the non-slot syntax.  This test should be extended as this is
just a start of something we can hopefully add to some
automated CI tests.

To run simply do:

./exercise_ledctl.py <ledctl binary> [optional slot filter N ]+

eg.

 # ./exercise_ledctl.py ../src/ledctl sg3- sg2-
 slot filter = ['sg3-', 'sg2-']
 # exit $?
 # 0

Only tested on SES hardware.  It *should* work with VMD and NPEM
too, but unable to test.  This test takes about 7 minutes to run
on our test setup.

Signed-off-by: Tony Asleson <tasleson@redhat.com>