ses/scsi: Update block_device->ibpi on set
Library users will expect that when they change a value of an LED that if
retrieve the value after the set that is will reflect its new value.
Additionally a library user should be able to seamlessly use the slot or
device path API if the LED hardware supports it and have the ibpi values
match.  This change allows SES hardware to meet these expectations.

Signed-off-by: Tony Asleson <tasleson@redhat.com>