Ledctl: do not check slot state for set locate_off
locate_off is used to reset the state of the led.
IBPI_PATTERN_LOCATE_OFF has no equivalent in the statuses of individuals
controllers and way of blinking, usually the status is set to some
equivalent of NORMAL, but it depends on the controller and mapping.

Checking if IBPI_PATTERN_LOCATE_OFF is already set on slot always ends
with unnecessary reading of the state/status of the physical led,
because reading result will be different from IBPI_PATTERN_LOCATE_OFF.

After this patch, locate_off will always be set,
which also happened before, but now without unnecessary checks.

Signed-off-by: Blazej Kucman <blazej.kucman@intel.com>