libled: API changes and removal
Change from bool to enum led_cntrl_type  for led_is_management_supported
which allows us to return not only if it is supported, but what
controller provides the preferred support for it.

Remove led_set as this doesn't work for all the different supported
hardware.  You need slot support to have the ability to retrieve the LED
status.

Signed-off-by: Tony Asleson <tasleson@redhat.com>