Nvme subsystem minimal support
To support multipath capable drives we need to know physical location
but use virtual devnode. This method should be sufficient for the drives
connected via one port.

Supporting dual-ports and full multipathing support (especially in
ledmon) is a problem that would require better implementation. This patch
makes it recognizable by ledctl and library in standard case (one port is
used).

Try to determine devnode on block_device_init and save it. We refer
to this property in led_device_name_lookup() now. It is fair to require
context for lookup.

It fixes get and set slot with multipath drives. Now we are comparing
request with devnode, so it removes possibility to use sysfs path in this
context.

Additionally, it fixes some verbose checkpatch errors and adds some other
improvements.

Co-authored-by: Tony Asleson <tasleson@redhat.com>
Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>