Nvme subsystem minimal support
To support multipath capable drives we need to get their physical location.
This method should be sufficient for the drives connected via one port.

Supporting dual-ports and full multipathing support (especially in
ledmon) is a problem that requires better implementation. This is low risk
effort to make it recognizable in ledctl and in library in standard cases.

Additionally, it fixes some verbose checkpatch errors and adds some other
improvements.

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>