libled: Specify public enum. values
Be specific and assign the value for each enumeration.
This prevents the case where a developer wants to add
a new entry and inserts in the middle of existing
enumerations.  If this happened, the compiler would
automatically re-assign different values to all the
enumerations, breaking existing library users at run-time.

This change also reserves some regions of numeric values
to ensure we have room for future additions.

Signed-off-by: Tony Asleson <tasleson@redhat.com>