add devnode for block device
Instead parsing and reparsing, try to determine devnode on
block_device_init and save it. We can refer to this property in
led_device_name_lookup(). It is fair to require context for lookup now.

It fixes get and set slot too. Now we can are comparing with devnode
but it removes possibility to use sysfs path in this context.

Signed-off-by: Mariusz Tkaczyk <mariusz.tkaczyk@linux.intel.com>