cntrl: remove regex from checking cntrl paths
ALLOWLIST and EXCLUDELIST set in ledmon.conf contain
paths to controllers and it was parsed using strcmp
or regex. Using regex caused unwanted behaviors, because
full paths to drives was accepted and user could block
controllers and doesn't be aware of it.
Now only strncmp is used to compare, and only precise
path to controller is accepted or "/sys/devices" to
block or allow all controllers.

Signed-off-by: Kinga Tanska <kinga.tanska@intel.com>