ledmon/amd: Add a new NVMe LED controlling method
  The purpose of this patch is to make AMD's NVMe led controlling
methods more diverse and meet the secondary development needs of
all customers as much as possible.
  The way of original ledmon to control NVMe led is:
  At present, there is only "out-of-band" NVMe control method, this
method also requires maintaining a mapping table and customers want
to control NVMe leds in a way similar to Intel VPP.
  For patch, a method is also proposed to control the PCIe slot crtl
register directly to control the NVMe Led. In this way, the customer
does not have the above limitation of controlling the NVMe led through
IPMI.

Signed-off-by: qiyuzhu <qiyuzhu2@amd.com>