ipmi.c: Fix off by 1 memset
The code was using the size of an element + number of elements
for the memset which is incorrect.  It should be the size of an
element * the number of elements.  However, as we are simply
clearing memory for an array on the stack, a simple sizeof
array will give us what we need, which is the number of bytes
the array occupies.

This was discovered as a SIGABORT on Fedora 38 testing of
pre-released compiled library package.

Note: replaced memsets for structs with struct initialization
syntax as requested in review comments.

Signed-off-by: Tony Asleson <tasleson@redhat.com>