Use inclusive nomenclature across the code
This patch replaces remainings of "master/slave" nomenclatrue.

Rename slave.c/h files to tail.c/h.
Replace occurrences of "slave" with "tail" where possible.

The "tail" was chosen due to it's abstract meaning.
Tail as in "to tail" is a  device that ledmon follows.

Some external libraries are still using old nomenclature;
thus, this code must be left in place.

Signed-off-by: Mateusz Kusiak <mateusz.kusiak@intel.com>