CI: move .checkpatch.conf file from main directory
When .checkpatch.conf is placed in main repo directory, it will be used
by system checkpatch when user will run scan on his patch. To avoid it,
and use this conf file only to GH action, it was moved outside main
repository. While GH action is running, conf file is being moved to main
directory back, which makes possible to use it while checkpatch action.

Signed-off-by: Kinga Stefaniuk <kinga.stefaniuk@intel.com>