docs: minor updates for accessibility
Several minor updates based on English grammar to help those with
low vision more easily read the docs.  For those with low vision,
like me, a missing word or misplaced article (a, the) can add to the
confusion while trying to read a sentence when the words tend
to 'jump around' a bit.

This is part of ledmon taking an ''accessibility pledge' to better
support diversity and inclusion.

Signed-off-by: Paul Luse <paul.e.luse@intel.com>