Add gcc 14 and fix GH actions (#240)
PR contains fixed to files which are using by GitHub actions.

Change runner to Ubuntu 24.04 which supports gcc versions up to 14.
Previously ubuntu-latest was used (22.04) which didn't support gcc 13
and 14. Add verification if correct gcc was installed during test.
Setting fetch-depth to 0 to fetch all of the history of all branches and
tags.
When .checkpatch.conf is placed in main repo directory, it will be used
by system checkpatch when user will run scan on his patch. To avoid it,
and use this conf file only to GH action, it was moved outside main
repository. While GH action is running, conf file is being moved to main
directory back, which makes possible to use it while checkpatch action.

---------

Signed-off-by: Kinga Stefaniuk <kinga.stefaniuk@intel.com>