Correct use of slot array index vs. SES index
The code incorrectly was using the array index to lookup the block device
and the SES slot instead of using the SES index value.  The hardware that
I have available to test has both of these values the same, so no issue
was found, but in any case where these don't match it would cause an
issue.

Variables and function parameters renamed to hopefully reduce confusion.

Signed-off-by: Tony Asleson <tasleson@redhat.com>